backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.3
asyncpg==0.29.0
python-dotenv==1.0.1
openpyxl==3.1.2
reportlab==4.0.9
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
